Fixed Bugs, Optimized Code, Removed Unnecessary Codes and many more...
I was going to add comments to this script so that in future I don't have to face problem to understand codes with my beloved forgetfulness. But unfortunately I already forgot almost everything in this script. So I decide to change variable named accordingly and add as much as comments possible to describe what line doing what in the script. I found some bugs middle of my work and fixed them as well. Whatever ultimately the changes I made here are,

 - [Added] Comments as much as possible to explain every line
 - [Fixed] 1 weak warning, 1 major bug & few minor bugs
 - [Changed] Variables name for better understanding
 - [Removed] unnecessary module, variables & lines of codes
 - [Reduced] Memory uses
 - Optimized codes

Signed-off-by: Ahsan40 <30434801+Ahsan40@users.noreply.github.com>